CentOS 7 gdb-rhbz1013453-value-struct-elt-memory-leak.patch: https://sourceware.org/ml/gdb-patches/2013-07/msg00469.html
https://sourceware.org/ml/gdb-patches/2013-07/msg00469.html

Hi.
This patch adds the missing calls to check_typedef
and adds a testcase to show the issue.

The PR is not closeable yet, but the remaining issues are more cleanups
than actual bug fixes.

Regression tested on amd64-linux.

I will check this in in a few days if there are no objections.

[The multiple calls to check_typedef (value_type (*argp)) are
bothersome, but the code is simpler this way, and I expect
resolving the rest of the issues in 15695 to potentially change this
code significantly.]

---

commit cbb25189b69e501ddca64917d810b54bb1466c93
Author: Doug Evans <dje@google.com>
Date:   Thu Aug 1 23:59:47 2013 +0000

    	PR symtab/15695
    	* valops.c (value_struct_elt): Add missing call to check_typedef.
    	(value_find_oload_method_list): Ditto.

    	testsuite/
    	* gdb.base/func-ptr.exp: New file.
    	* gdb.base/func-ptr.c: New file.

2013-08-01  Doug Evans  <dje@google.com>

	PR symtab/15695
	* valops.c (value_struct_elt): Add missing call to check_typedef.
	(value_find_oload_method_list): Ditto.

2013-08-01  Doug Evans  <dje@google.com>

	PR symtab/15695
	* gdb.base/func-ptr.exp: New file.
	* gdb.base/func-ptr.c: New file.