CentOS 7 gdb-rhbz1105165-ibm-tdb-support-system-z-4of9.patch: [PATCH] S390: Fix TDB regset recognition
From 6682d9595ed8d9b9bba5470bfd7fd1ccd378f19a Mon Sep 17 00:00:00 2001
From: Andreas Arnez <arnez at linux dot vnet dot ibm dot com>
Date: Tue, 5 Nov 2013 18:43:50 +0100
Subject: [PATCH] S390: Fix TDB regset recognition

When checking for the presence of the TDB regset, the current code
interprets ENODATA from PTRACE_GETREGSET as an indication that the TDB
regset *could* occur on this system, but the inferior stopped outside
a transaction.  However, the Linux kernel actually reports ENODATA
even on systems without the transactional execution facility.  Thus
the logic is now changed to check the TE field in the HWCAP as well.

This version also checks the existence of the TDB regset -- just to be
on the safe side when running on TE-enabled hardware with a kernel
that does not offer the TDB regset for some reason.

gdb/
	* s390-linux-nat.c (s390_read_description): Consider the TE field
	in the HWCAP for determining 'have_regset_tdb'.

gdbserver/
	* linux-s390-low.c (HWCAP_S390_TE): New define.
	(s390_arch_setup): Consider the TE field in the HWCAP for
	determining 'have_regset_tdb'.
---
 gdb/ChangeLog                  |  5 +++++
 gdb/gdbserver/ChangeLog        |  6 ++++++
 gdb/gdbserver/linux-s390-low.c | 48 ++++++++++++++++++++++++++----------------
 gdb/s390-linux-nat.c           |  5 +++--
 4 files changed, 44 insertions(+), 20 deletions(-)