CentOS 7 gdb-rhbz809179-global-variable-nested-dso.patch: [PATCH] Add test for global variable that is nested by another DSO
This bug is already fixed in the RHEL-7.1 GDB version.  However, it is
also important to add a testcase, so the following is a patch to add
this test.

This is the original message submitted to gdb-patches:

From: Sergio Durigan Junior <sergiodj at redhat dot com>
To: GDB Patches <gdb-patches at sourceware dot org>
Subject: [PATCH] Add test for global variable that is nested by another DSO
Date: Thu,  4 Sep 2014 17:32:53 -0400
Message-Id: <1409866373-16413-1-git-send-email-sergiodj@redhat.com>

This is just a testcase addition that I am proposing for upstream GDB.
We have this in our internal tree, and the related RH bug is:

  <https://bugzilla.redhat.com/show_bug.cgi?id=809179>

(You might not be able to see all the comments without privileges.)

This bug is about a global variable that got incorrectly displayed by
GDB.  This bug has already been fixed a long time ago by Joel's
commit:

  commit 19630284f570790ebf6d50bfb43caa1f125ee88a
  Author: Joel Brobecker <brobecker@gnat.com>
  Date:   Tue Jun 5 13:50:50 2012 +0000

But I think a testcase for it wouldn't hurt.

So, consider the following scenario:

  $ cat solib1.c
  int test;
  void c_main (void)
  {
    test = 42;
  }

  $ cat solib2.c
  int test;
  void b_main (void)
  {
    test = 42;
  }

  $ cat main.c
  int main (int argc, char *argv[])
  {
    c_main ();
    b_main ();
    return 0;
  }

  $ gcc -g -fPIC -shared -o libSO1.so -c solib1.c
  $ gcc -g -fPIC -shared -o libSO2.so -c solib2.c
  $ gcc -g -o main -L$PWD -lSO1 -lSO2 main.c
  $ LD_LIBRARY_PATH=. gdb -q -batch -ex 'b c_main' -ex r -ex n -ex 'p test' ./main
  ...
  $1 = 0

This happened with GDB before Joel's commit above.  Now, things work
and GDB is able to correctly display the nested global variable:

  $ LD_LIBRARY_PATH=. gdb -q -batch -ex 'b c_main' -ex r -ex n -ex 'p test' ./main
  ...
  $1 = 42

The testcase attached tests this behavior.

gdb/testsuite/ChangeLog:
2014-09-04  Sergio Durigan Junior  <sergiodj@redhat.com>

    * gdb.base/global-var-nested-by-dso-solib1.c: New file.
    * gdb.base/global-var-nested-by-dso-solib2.c: Likewise.
    * gdb.base/global-var-nested-by-dso.c: Likewise.
    * gdb.base/global-var-nested-by-dso.exp: Likewise.